+++ apps/backend/src/modules/route-categories/libs/types/types.ts
@@ -1,4 +1,0 @@
-export {
-	type RouteCategoryGetAllItemResponseDto,
-	type RouteCategoryRequestDto,
-} from "@smartscapes/shared";