+++ apps/frontend/src/libs/types/types.ts
@@ -1,5 +1,6 @@
 export { type AsyncThunkConfig } from "./async-thunk-config.type.js";
 export { type IconName } from "./icon-name.type.js";
+export { type Location } from "./location.type.js";
 export { type NavigationItem } from "./navigation-item.type.js";
 export { type NavigationItemsGroup } from "./navigation-items-group.type.js";
 export {
@@ -7,6 +8,8 @@ export {
 	type APIResponse,
 	type APIValidationErrorDetail,
 	type Coordinates,
+	type PaginationMeta,
+	type PointGeometry,
 	type ValidationSchema,
 	type ValueOf,
 } from "@smartscapes/shared";