+++ apps/frontend/src/modules/routes/libs/types/types.ts
@@ -1,5 +1,6 @@
 export { type PatchActionPayload } from "./patch-action-payload.type.js";
 export {
+	type RouteFindAllOptions,
 	type RouteGetByIdResponseDto,
 	type RoutePatchRequestDto,
 } from "@smartscapes/shared";