+++ apps/frontend/src/modules/routes/slices/routes.ts
@@ -1,8 +1,9 @@
-import { getRouteById, patchRoute } from "./actions.js";
+import { getAll, getRouteById, patchRoute } from "./actions.js";
 import { actions } from "./route.slice.js";
 
 const allActions = {
 	...actions,
+	getAll,
 	getRouteById,
 	patchRoute,
 };