+++ packages/shared/src/index.ts
@@ -9,9 +9,10 @@ export {
 	AppEnvironment,
 	CommonExceptionMessage,
 	ContentType,
+	LocationType,
 } from "./libs/enums/enums.js";
 export { ValidationError } from "./libs/exceptions/exceptions.js";
-export { configureString } from "./libs/helpers/helpers.js";
+export { changeStringCase, configureString } from "./libs/helpers/helpers.js";
 export { type Config } from "./libs/modules/config/config.js";
 export {
 	type HTTP,
@@ -29,11 +30,20 @@ export {
 	type APIValidationErrorDetail,
 	type Coordinates,
 	type LineStringGeometry,
+	type PaginationMeta,
+	type PaginationQuery,
 	type PointGeometry,
 	type ValidationSchema,
 	type ValueOf,
 } from "./libs/types/types.js";
 export { AuthApiPath, AuthExceptionMessage } from "./modules/auth/auth.js";
+export {
+	CategoriesApiPath,
+	type CategoryCreateRequestDto,
+	categoryCreateValidationSchema,
+	CategoryExceptionMessage,
+	type CategoryGetAllItemResponseDto,
+} from "./modules/categories/categories.js";
 export {
 	FileFolderName,
 	type FileMimeType,
@@ -62,13 +72,15 @@ export {
 	PermissionKey,
 } from "./modules/permission/permission.js";
 export {
-	pointOfInterestCreateValidationSchema,
 	pointOfInterestUpdateValidationSchema,
 	PointsOfInterestApiPath,
+	pointsOfInterestCreateValidationSchema,
+	type PointsOfInterestPaginatedResponseDto,
+	type PointsOfInterestPaginatedSummary,
+	type PointsOfInterestQueryRequest,
+	pointsOfInterestQueryValidationSchema,
 	type PointsOfInterestRequestDto,
 	type PointsOfInterestResponseDto,
-	type PointsOfInterestSearchQuery,
-	pointsOfInterestSearchQueryValidationSchema,
 } from "./modules/points-of-interest/points-of-interest.js";
 export {
 	type ReviewCreatePayload,
@@ -76,12 +88,6 @@ export {
 	type ReviewGetByIdResponseDto,
 	type ReviewRequestDto,
 } from "./modules/reviews/reviews.js";
-export {
-	RouteCategoriesApiPath,
-	RouteCategoryExceptionMessage,
-	type RouteCategoryGetAllItemResponseDto,
-	type RouteCategoryRequestDto,
-} from "./modules/route-categories/route-categories.js";
 export {
 	type RouteConstructRequestDto,
 	type RouteConstructResponseDto,