+++ packages/shared/src/libs/helpers/change-string-case/change-string-case.helper.ts
@@ -1,0 +1,7 @@
+import * as changeCase from "change-case";
+
+const changeStringCase = (string_: string): string => {
+	return changeCase.camelCase(string_);
+};
+
+export { changeStringCase };