+++ packages/shared/src/modules/points-of-interest/libs/validation-schemas/points-of-interest-query.validation-schema.ts
@@ -1,0 +1,115 @@
+import { z } from "zod";
+
+import { parseToFloat } from "../../../../libs/helpers/helpers.js";
+import {
+	PointsOfInterestValidationMessage,
+	PointsOfInterestValidationRule,
+} from "../enums/enums.js";
+
+const pointsOfInterestQuery = z
+	.object({
+		latitude: z
+			.string()
+			.trim()
+			.transform(parseToFloat)
+			.pipe(
+				z
+					.number()
+					.min(
+						PointsOfInterestValidationRule.LATITUDE_MIN,
+						PointsOfInterestValidationMessage.LATITUDE_MIN,
+					)
+					.max(
+						PointsOfInterestValidationRule.LATITUDE_MAX,
+						PointsOfInterestValidationMessage.LATITUDE_MAX,
+					),
+			)
+			.optional(),
+		longitude: z
+			.string()
+			.trim()
+			.transform(parseToFloat)
+			.pipe(
+				z
+					.number()
+					.min(
+						PointsOfInterestValidationRule.LONGITUDE_MIN,
+						PointsOfInterestValidationMessage.LONGITUDE_MIN,
+					)
+					.max(
+						PointsOfInterestValidationRule.LONGITUDE_MAX,
+						PointsOfInterestValidationMessage.LONGITUDE_MAX,
+					),
+			)
+			.optional(),
+		name: z
+			.string()
+			.trim()
+			.min(
+				PointsOfInterestValidationRule.NAME_MIN_LENGTH,
+				PointsOfInterestValidationMessage.NAME_MINIMUM_LENGTH,
+			)
+			.max(
+				PointsOfInterestValidationRule.NAME_MAX_LENGTH,
+				PointsOfInterestValidationMessage.NAME_MAXIMUM_LENGTH,
+			)
+			.optional(),
+		page: z
+			.string()
+			.trim()
+			.transform(parseToFloat)
+			.pipe(
+				z
+					.number()
+					.min(
+						PointsOfInterestValidationRule.MIN_PAGE,
+						PointsOfInterestValidationMessage.PAGE_MUST_BE_GREATER_THAN_ZERO,
+					),
+			)
+			.optional(),
+		perPage: z
+			.string()
+			.trim()
+			.transform(parseToFloat)
+			.pipe(
+				z
+					.number()
+					.min(
+						PointsOfInterestValidationRule.MIN_PER_PAGE,
+						PointsOfInterestValidationMessage.PER_PAGE_MUST_BE_GREATER_THAN_ZERO,
+					),
+			)
+			.optional(),
+		radius: z
+			.string()
+			.transform(parseToFloat)
+			.pipe(
+				z
+					.number()
+					.min(
+						PointsOfInterestValidationRule.RADIUS_MIN_KM,
+						PointsOfInterestValidationMessage.RADIUS_MIN_KM,
+					)
+					.max(
+						PointsOfInterestValidationRule.RADIUS_MAX_KM,
+						PointsOfInterestValidationMessage.RADIUS_MAX_KM,
+					),
+			)
+			.optional(),
+		search: z.string().trim().optional(),
+	})
+	.refine(
+		({ latitude, longitude, radius }) =>
+			!radius || (Boolean(latitude) && Boolean(longitude)),
+		{
+			message: PointsOfInterestValidationMessage.COORDS_REQUIRED_WITH_RADIUS,
+		},
+	)
+	.refine(
+		({ latitude, longitude }) => Boolean(latitude) === Boolean(longitude),
+		{
+			message: PointsOfInterestValidationMessage.COORDS_REQUIRED_TOGETHER,
+		},
+	);
+
+export { pointsOfInterestQuery };